Move configuration definitions to separate file.
